requirements.txt
python-telegram-bot==20.3
Flask==2.3.2
psycopg2-binary==2.9.6
gunicorn==20.1.0